fix: autotests errors

+++ migrations/000001_create_metrics_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE metrics (
     id VARCHAR(255) NOT NULL,
     type VARCHAR(7) NOT NULL,
-    delta INT NULL,
+    delta BIGINT NULL,
     value DOUBLE PRECISION NULL,
     hash VARCHAR(511) NULL,
     PRIMARY KEY(id, type)
